Allow different levels for debug traces.
For future additions to these scripts, there are some traces that
are only useful some times, and are otherwise a potential source
of a lot of noice (traces generated for each change in a given commit,
for instance).  For this purpose, we introduce the notion of debug
level.

To reflect this change, the name of the environment variable and
git config variable has been changed to HOOKS_DEBUG_LEVEL and
hooks.debuglevel accordingly.