Rename bare-repo.git repos into bare/repo.git in all testcases.
The reason for this change is that the bare repo's directory name is
used to determine the name of the project. In our testcases, it looks
nicer if the project name is "repo" rather than "bare-repo".

Another change that snuck-in is the fact that we're now using symbolic
links to install the hooks, rather than copying them.