Use tarballs to store the testcases' git repositories.
The current approach was not working, because it is hard to check in
a complete bare repository.  Also, checking git repositories had the
disadvantage  that the git repositories always contain a LOT of files,
flooding the other changes with those changes. It is easier in all
counts to just use an archive.

This does not fix the testcases themselves, but the checked in versions
do not work anyway, it doesn't make things worse in practice.