Change a bit the AbstractRefChange interface.
Previous, it required that the child classes override the __init__
method, but this has the drawback of also forcing the child classes
to call the parent's init.

The purpose of asking the child classes to override __init__ was
to set a couple of attributes.  Instead, set them by calling two new
abstract methods.

Adjust the LightweightTagCreation class accordingly.