Code re-organization introducing "updates" module and submodules.
The goal is to move everything related to a given type of update
to its own file, rather than aggregate support for every kind of
update inside the same file.

This will also allow us to aggregate handling of update-validation
and email sending, with the view of sharing the code that computes
the list of commits to be emailing about.