Add verification that our scripts do not leak any temporary file.
This patch forces the scripts to use a specific directory as the root
for their own temporary directories/files. We create it as an empty
directory at the start of the testcase, and then verify at the end
that it is still empty. If not, the testcase will report a failure.

This patch also makes sure that, even if the scripts do leak some
files/directories, the testsuite catches those and deletes them.