Enable pre-commit checks for all types of updates.
For instance, if a push updates a tag to point to a commit that
is not referenced by any branch yet, then run the pre-commit check
on all new commits introduced by this push (those that are not
accessible from any branch).

Move expand_new_commit_to_list from the updates.branches module
to the updates modules, since this is no longer branch-specific.

Adjust the rest of the code, as well as some of the testcases
accordingly.