Sanitize output from git commands run from testcase.
There is a trailing \033[K sequence (terminal control sequence
that tells the terminal to erase the end of the line) which seems
to be generated by git, and it makes matching the output a little
more difficult than necessary.  So we create a subclass of
gnatpython.ex.Run which provides an additional property containing
the santized output.

While at it, we also added another property which replaces the
ex_run_image function.

Update all testcases throughout.