Get rid of _quiet option for git_run.
For our purposes, we would always want _quiet to be True, whereas
the default was False, leading to a lot of situations where we forgot
to set it.  So we delete the option.  To access the command's output
in case of command failure, now use the CalledProcessError exception's
output method.

One of the changes also introduced by this patch is the fact that
stderr is now redirected to stdout.

This will avoid seeing some inexplicable spaces in the output
printed by git push commands. These spaces were in fact coming
from the "print" statements that we just got rid of - triggered
by the fact that some calls to git_run were missing the _quiet
option.