no_mailinglist_update: Full and exact matching of expected output.
The XFAIL has also been removed in favor of just recording the current
output. The failure is a temporary missing feature while we re-implement
the git-hooks, rather than an actual bug.  In the meantime, any change
outside of this will also be seen, whereas before, it would have still
been marked XFAIL.