Add Updates.pre_update_refs attribute.
This attribute is used to store the value of all references
before an update is applied.

This was computed on-demand by expand_new_commit_to_list, but we want
to make this a parameter of this function, in order to allow us to
call this function from any hook.  In particular, in the post-receive
hook, all reference values have already been updated, so the only
way to get the value of a reference prior to the update is by reverse-
applying the ref updates passed by argument to the post-receive hook.

The ultimate goal of this patch is to re-use expand_new_commit_to_list
during the phase where we send emails for each new commit.