AbstractUpdate: Add added_commits/lost_commits attributes...
... implemented as properties.  This allows us to simplify a bit
several of the methods defined by this class.

This will also make this information available for the methods
which were not passed these lists. This should become useful
for methods such as NotesUpdate.validate, which now has access
to the list of notes being updated, and thus can verify that
the annotated commit exists in the repository.

Part of LC24-002.