New method support.TestCase.assertRunOutputEqual
This method verifies that the output of a Run is equal to the given
expected output, and otherwise prints the command's image, actual
output, followed by a diff. The "diff" is the part that's new, and
allowed to quickly determine the difference, as the expected output
tends to be quite long in a typical test.

Update all testcases accordingly.