Reimplement config.py to call "git config" only once.
This patch currently removes the 'type' info from the GIT_CONFIG_OPTS
objects, but this will be put back later in a slightly more powerful
form (ie: the module will be using this type info to convert the
config option into the right type, instead of just passing it to
"git config" to validate the value).

For LC27-007.