Add support for returning typed value by config.git_config.
This patch allows us to specify a given type to any git config
option, and have git_config config the option value (a string)
into that type, raising an InvalidUpdate error if the conversion
failed.

Currently no such option is taking advantage of this features.
Transition of the relevant options will be done piecemeal.

Part of LC27-007.