Email: Move the creation of the MIMEText object from __init__ to send.
This is preparation work for LC28-010, where we want to introduce
a delay between sending each email, but without having to wait
ourselves. This means that the emails will have to be created,
and then queued.  It's not clear whether this makes any difference,
but this allows us to make sure that the creation of the MIMEText,
and possibly associated fields such as the "Date:", are correlated
to the actual moment the email is sent.