Implement an email queue
The hooks now enqueue all emails in that queue, and post_receive_email
now triggers the flushing of that queue at the end of the script
execution.  The emails are still sent within the same process, but
the goal is to daemonize the part that sends it.

Part of LC28-010.