push_submodule_creation: Fix spurious error on kani.
Kani is a machine where the cvs_check program is installed at
the standard location, and thus gets called by the testcase.
This is a bit of an accident, but isn't a bad accident. So adjust
the testcase's fake cvs_check program to generate the same output
as the real one, and adjust the expected output accordingly.

This way, the testcase can be run on both types of machines (with
or without the real cvs_check installed).