New attribute AbstractUpdate.email_info
The primary motivation behind this patch is simplification of
the code, after having noticed that both the update and post-receive
hooks create the object.  So we can make it an attribute computed
at instantiation time, and simplify many of the methods implemented
by this class.

The second motivation behind this patch is also a slight performance
improvement, measurable when the number of files changed by a push
is getting more important. This is the change in pre_commit_checks.py,
which now takes an extra argument providing the project/repo name,
instead of recomputing it (involving running git) for each and every
file being checked.

The performance gain is fairly modest, measured to be around 4ms per
file being checked, but worth taking, given the fact that it does not
complicate the code in pre_commit_checks.