commit_subject: Only request one commit in call to "git rev-list"...
... instead of getting the entire list of commits, only to keep
the first one.  This dramatically improves performance in repositories
such as GDB where branches contain a very large number of commits.
Combined with the identical type of fix done in commit_oneline,
the time required for a "git push" in the GDB repository went down
from 40 seconds down to 3.5 seconds!