Revert "Make sure the info/attributes file is group-writable."
It turns out that making info/attributes group-writable is not
sufficient: A second user is allowed to overwrite the file, but
then the hooks fail when trying to change the file permissions
to make it group-writable.

We will use a different approach which should not need this anymore.

Part of M116-004.