Implement filename-collision checks.
This implements an additional check, to be applied for every new commit,
which will reject a commit if the check detects that two files have
the same name modulo casing.  This will help make the project nicer
to the users when Operation Systems such as Darwin or Windows, where
filesystems are case-insensitive, are involved.

For M524-010.