wrong commit use as reference for commit-per-commit pre-commit-check.
This patch should fix problems occurring when pushing merge commits.
See large comment in updates.__added_commits for more info.

Update all relevant testcases. And add a couple of testcases, one
in the usual mode, and one in combined-style-checking mode. Note
that the addition of the new testcases was not strictly necessary,
since several older testcase had clues showing a related issue
as well. For instance, merge_commit_branch_update shows that we had
a testcase that showed we were style-checking the wrong file.
But the new testcases are nice to have anyway, as it also allows us
to test more specifically for the situation that triggered this change.

For M606-034.