Avoid triggering ticket filing of "cover email".
This patch mostly removes FILER_EMAIL from the list of recipients
of the "cover email". It still keeps it for annotated tags, though,
for reasons explained as comments in the associated code.  To prevent
the "summary of changes" part of the email from triggering the filing,
add a "Diff:" marker prior to the summary-of-changes.

Update the testsuite accordingly.  A new testcase is not necessary
as this is already thoroughly exercised both ways, as shown by
the extensive number of adjustements required in the testsuite
following this change.

For M612-024.