Call the style-checker with the relative path to the file to be checked.
The style-checker is now called a bit differently. Instead of copying
the file to be checked straight into our temporary directory, and then
calling the style-checker with an absolute path (which means nothing
to the style-checker), we now do the following:

  - Recreate the directory hierarchy where the file lives inside
    our temporary directory, copy the file to be checked there;

  - Call the style-checker with the CWD set to the temporary directory
    and with a relative path to the file being checked as the second
    argument.

This makes it possible for the style-checker to provide error messages
citing the path to the file, making it:

  - Easier for the user to locate the file in question;

  - Possible for the user to determine which file is being checked,
    when the same filename is used multiple times across the project
    (typical of testsuites, for instance).

For MA09-043.