Avoid use of "cvs_check" in favor of "style-checker" wherever possible.
Since we're about to allow the use of an alternate style-checker,
and since the name "cvs_check" contains the name of an old technology
we abandonned many years ago, it seems like a good idea to introduce
a more neutral name for that tool.

There is one, non-upward-compatible user-visible change: The script
now uses the GIT_HOOKS_STYLE_CHECKER environment variable, instead
of the GIT_HOOKS_CVS_CHECK environement as a backdoor to specifying
an alternate style-checker.  This should be OK, however, as this
feature was really only meant to be used during testing, and is,
in practice, only really usable in that context.

This is prep work before introducing a new config option allowing
the repository to use an alternate style checker.

Part of MA09-043.