Add new config parameter hooks.max-rh-line-length.
This new parameter allows each project to decide which maximum
line length to authorize in their revision logs. It also allows
each project to turn this check off entirely, by setting its
value to zero.

The default behavior remains unchanged, which is a maximum line
length of 76 characters.

For M823-001.