update_l_tag: Add --force flag to "git push".  Adjust expected output.
This is to support newer versions of git, which now reject by default
tag _updates_ (ie, pushing new values for an already-existing tag).
Unfortunately, this also triggers a difference in output compared to
older versions, so some massaging on the actual output has been added
to transform the old output into the new one, pretending that all
versions of git produce the new output.