update_tag: Use --force to push new tag value.
This is required by newer versions of git, which now rejects by default
pushes that update the value of already-existing tags.

The new version also introduces a difference in output, which is
taken into account in the new expected output.  But this means that
we need to add some extra output massaging to handle the output from
older versions of git.