Create project.config in refs/meta/config branch (from hooks_config).
We are still preserving the hooks_config file within the test
directory in order to help grep-ability, but this is no longer
the master.  The pack-test-repos script has been updated to grab
and update our convenience copy to match the latest master.

Similarly, the unpack-test-repos script has been updated to use
the project.config file (ie, use the master copy) to populate
the bare repository's config file. It is expected that this part
will become unnecessary at some point, when the hooks are enhanced
to get their config from this file, instead of the repository's
config file.

For MC20-031.