Make it explicit that branch deletions in refs/meta are not allowed...
Maybe there will be a need to allow certain branches in refs/meta
to be deletable, but for now, that namespace has a very specialized
use and while we haven't had a situation were deleting a branch in
that namespace has been needed, we certainly don't want the "config"
branch to be ever deleted, since we rely on it to get the hooks'
configuration. So, keep the current situation, which is the easiest
to implement: disallow all branch deletions in that namespace.

This patch just makes that decision explicit.