Re-implement file locking using hard links instead of sockets.
Not sure why, but trying to create sockets on some machines triggered
an error, maybe because of insufficient privileges or unportable
socket usage. As a result, all pushes were systematically rejected,
with the hooks thinking that they had detected a concurrent push
request.

This patch re-implement the locking mechanism using a filesystem-based
approach (hard links).

For M903-031.