Improve handling of multiple email addresses in hooks.mailinglist
The current code already handles properly hooks.mailinglist configs
that contain multiple email addresses. This patch slightly enhances it
make the hooks.mailinglist config a "tuple" instead of a simple string.

This will prepare the way for handling entries that are scripts
returning email addresses, which is going to be useful in the context
of the binutils-gdb.git project, where the list of recipient is
dependent on the files being changed by the commit.

Note that this feature is already being tested by the "multiple_commits"
test, so an additional test is not necessary.

For NC06-002.