Preparation work for "cover email" dynamic distribution list.
This patch enhances the get_update_email_contents method of all
child classes of AbstractUpdate which were returning email
subject and body to also return a distribution list. For now,
all cover emails are always sent to everyone, as this is what
seems to make the most sense.

At the moment, the distribution list is still completely static,
but, for cover emails, the computation of that list is centralized
in a new AbstractUpdate.everyone_emails method, which will be
enhanced later to handle scripts as well.

For NC07-001.