Do not aggregate email_body and diff in class Email until needed.
This change slightly re-organizes the Email class to split the
email_body attribute back into email_body (pre-existing, obviously)
and diff (new attribute), thus preserving the email_body and diff
parameters passed during instantiation.  It then re-combined them
into the full email body just before the email is to be sent,
thus making this change a no-op from the user's perspective.

This is preparation work for allowing the call of an external script
to which we would be passing the email_body without the contents
of the diff.

For NC20-011.