post-receive, update: Add hardcoded path to Git & Python.
Unfortunately, both git and Python are too old on sourceware.org, so
a local install has been made as a workaround, and this patch updates
the PATH to use them from the two scripts that our hooks define.

Note that this is not completely exhaustive, since other operations
get performed by git outside of hooks, and those can still cause
trouble during the testsuite run. But, for now, things seem to be
relatively in control, as the number of failures associated to that
are fairly small, and can be controlled by hand.