cannot push non-merge commit on branch where merge commits are rejected.
Trying to push a simple non-merge commit on a branch for which
hooks.reject-merge-commits matches fails, with the hooks rejecting
the update saying that merge commits are not allowed.

This is due to an obvious bug in reject_commit_if_merge which
tests the number of parents of a commit and causes an error
if the number of parents is at least 1. But the correct test
is for the number of parents to be at least 2 (that's the definition
of a merge commit).

Fixes NC27-001.