AbstractUpdate: rename __set_commits_attr into __set_send_email_p_attr
This method is only used to set the send_email_p attribute for commits
which don't exist in any of the branches matches the hooks.no-emails
config. Instead of keeping this method general-purpose, make it specific
to this specific need.  This is preparation work, before adjusting
the filtering performed by this method in a way that would be even more
specific to the handling of the hooks.no-emails config.

For P329-007.