remove syslog when style-check disabled by .gitattributes entry
Before this patch, we were generating one syslog entry for each
and every file that had the no-precommit-check attribute set -
meaning no style violation cvs_check. This not only generated
a lot of syslog entries, but also generated them for situations
that are not error situations. Since it is easy to determine
that information at any time in time, having that syslog trace
is really not necessary.

For P524-001.