remove syslogging for updates on branches with hooks.no-precommit-check
For updates on branches for which the hooks.no-precommit-check hook
matches the branch name, these hooks have been syslogging the fact
that all checks have been disabled due to that repository configuration.
Given that this option is routinely used (for managing branches that
are maintained by third-parties, for instance, where we may have no
control), and given that it is easy to determine that anytime after
the fact, there is no point in keeping that syslog entry. This patch
removes it.

For P524-001.