ignore temporary references created by gerrit during web edition
This patch adds another entry to the list of references to be ignored
(in GERRIT_INTERNAL_REFS). These are the temporary references which
gerrit creates when editing a commit directly from the gerrit web
interface. These are not user-visible references, and therefore
should not influence the behavior of our hooks.

For P812-021.