minor enhancement in cvs_check script used in various testcases
Given the recent change where we're preparing for calling the style_checker
only once for all files at once, the hooks are no longer printing
the name of the file that failed style checking. For that, we rely
on the style_checker itself. What this commit does is adjusting
the fake cvs_check scripts we use as the style checker within
this testsuite to mimick the desired behavior a little more.
Rather than just printing an error message, we print the name
of the file that triggered the error as well.

Related to Q530-009.