show correct location when identifying non-ISO-8859-15 characters in rev log
There was a bug when returning an error message indicating that
the revision contains non-ISO-8859-15, that would always show
the first line has having the unwanted character, even when
that characters was actually on a different line.

And while at it, I improved the error message to include the line
number as well. Just in case showing the text itself is either not
sufficient or just not as efficient.

Credit for finding the bug: Thomas Quinot (thank you!).

Followup on QB25-001.