Replace README file by README.md file (imported as is from GitHub version)
We made the effort of writing some nice documentation in the GitHub
version of the git-hooks; might as well import that documentation
in the AdaCore version. There is one minor section (about file-ci)
which clearly speaks to the GitHub users, it's really clear this
does not apply to this version. Since we're thinking of adjusting
the handling of this feature in the git-hooks so as to find a way
to reconcile this version of the hooks with the version published
on GitHub, it's just easier to keep the same doc for both.