Avoid uncovered line of code in CommitInfo.base_rev_for_display
This line of code was highlighted as uncovered by the testsuite's
coverage analysis, and reveals that we never call this method with
a CommitInfo object whose `parent_revs' attribute is None.

This attribute is first set during the class's initialization
(__init__) to the value of the parameter of the same name. And
the reason why we check for None here is because the __init__
routine documents this paramenter as allowing None, signifying
that the actual value hasn't been computed yet. The purpose of
this is to allow users who know that they don't need the parents
to skip computing them just for the sake of creating an object
of this class.

Upon investigation, it turns out that the only place where objects
of this class are instantiated is in function commit_info_list, where
the list of parents is actually computed and provided to the class.
This shows that, as of today, this uncovered code is actually
unreachable.

There were several options:
  (a) It is conceivable that some callers might find it useful
      that the parents get automagically computed on demand,
      arguing in favor of keeping that code. If we keep the code,
      we can cover it via Unit Testing.
  (b) Document that the instantiators of this class must always
      provide the list of parents at instantiation, thus making
      the check for None moot, allowing us to simply remove it.
  (c) Continue to allow the use of None for the list of parents,
      but still removing the code, transfering the responsibility
      to the users to be sure that it is not None before calling it.

In the end, it felt better to follow what is true today rather than
what "might be" tomorrow, so this commit rejects option (a).
Option (b) is simplest, but has the drawback of reducing flexibility
in terms of performance tweaking.  This is why this commit chose
option (c), especially since it's trivial for users to handle this,
as it's just a matter of calling one function.

Change-Id: I3a01ab52c06ae36f249e0c01fbc7ea59616e6ded