Reject early attempts to delete refs/meta/config
This is a followup on T227-003, where we enhanced the hooks to take
refs/meta/config updates into account right away. A small side-effect
of that change is that the error message when trying to delete the
refs/meta/config reference changed from...

    | This type of update (refs/meta/config,delete) is currently unsupported

... to ...

    | Unable to find the file project.config in refs/meta/config.
    |
    | Your repository appears to be incorrectly set up. Please contact
    | your repository's administrator to set your project.config file up.

The former wasn't all that great, but understandable. The new one
is not incorrect, but not very helpful. This patch therefore enhances
the hooks to provide a clear error message:

    | Deleting the reference refs/meta/config is not allowed.
    |
    | This reference provides important configuration information
    | and thus must not be deleted.

A new testcase is added to make sure we continue verifying
the error message in that situation.

Change-Id: I5178b40c0d003fbf8ffeb456592750ec38a9a047