Add handling of case where update's old_rev and new_rev are both null
Up until recently, it was thought that it was not possible for the hooks
to be called given a reference name with an old_rev and a new_rev
both null. As it happens, it actually is possible, and so this commit
adds better handling of this situation (instead of failing on an
assertion that they aren't both null).

A couple of testcases already testing the deletion of branches and
tags have been expanded to try the same, but using a full reference
name, instead of just the branch or tag's name. There was no need
to add a test for git notes, because there is no "short" reference
name in this case. So the test verifying git notes deletion was already
using the full reference name...

A test has been added to verify the behavior in the case when a full
reference name is used, and that reference does not exist in the remote.

Change-Id: If6e657312611ecbb0a66d0a8ae3763379d97e96f
TN: T329-003