human_readable_ref_name: do not ignore default_namespace parameter
This commit fixes an oversight in the implementation of the
AbstractUpdate.human_readable_ref_name method, where the
default_namespace parameter was unintentionally ignore because
a hardcoded value was used instead.

This currently has no visible effect for now, because the only references
for which we have been using this method are currently always branches.
This will make a difference as soon as we use this method for other
types of references, such as tags for instance.

Change-Id: I94fe39523dd591f387990aa6884f06c993af1c1c
TN: T209-003